pkg/endpoint: move FilterEPDir to directory.go
Signed-off by: Ian Vernon <ian@cilium.io>
