+++ pkg/endpoint/directory.go
@@ -22,3 +22,16 @@ import (
 func (e *Endpoint) directoryPath() string {
 	return filepath.Join(".", fmt.Sprintf("%d", e.ID))
 }
+
+// FilterEPDir returns a list of directories' names that possible belong to an endpoint.
+func FilterEPDir(dirFiles []os.FileInfo) []string {
+	eptsID := []string{}
+	for _, file := range dirFiles {
+		if file.IsDir() {
+			if _, err := strconv.ParseUint(file.Name(), 10, 16); err == nil {
+				eptsID = append(eptsID, file.Name())
+			}
+		}
+	}
+	return eptsID
+}
